requirements.txt
numpy>=1.26,<3.0
pillow>=10.0,<12.0
torch>=2.2,<3.0
torchvision>=0.17,<1.0
faiss-cpu>=1.8,<2.0
pyarrow>=15.0,<20.0
tensorboard>=2.16,<3.0
setuptools<81
transformers>=4.40,<5.0
datasets>=4.0,<5.0
tqdm>=4.66,<5.0
pytest>=8.0,<9.0
open_clip_torch>=2.24,<3.0
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0.0
pymongo[srv]>=4.6.0
fsspec>=2024.2.0
aiohttp>=3.9.0
requests>=2.31.0